engine/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.0
psutil>=5.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
